extract image-area.js
