+++ src/index.html
@@ -7,7 +7,8 @@
     <div id='sample-lazy-load'>
     </div>
 
-    <div id='sample-content'></div>
+    <div id='sample-content'>
+    </div>
 
   </body>
 </html>
